backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
python-dotenv==1.0.1
bcrypt==4.2.0
